fixed issue with sentiment score
This fixes an issue with the Json serializer and the sentiment
analysis. If a users first message had a weighting of 0, score would be
divided by 0.

Additionally, Gson would throw an exception when attempting to
serialize NaN. For this reason, an additional check was added to ensure
that the score and weighting remain a number.